move jobs to procs
