+++ xonsh/aliases.py
@@ -20,7 +20,7 @@ from xonsh.cli_utils import Annotated, Arg, ArgParserAlias
 from xonsh.dirstack import _get_cwd, cd, dirs, popd, pushd
 from xonsh.environ import locate_binary, make_args_env
 from xonsh.foreign_shells import foreign_shell_data
-from xonsh.jobs import bg, clean_jobs, disown, fg, jobs
+from xonsh.procs.jobs import bg, clean_jobs, disown, fg, jobs
 from xonsh.lazyasd import lazyobject
 from xonsh.platform import (
     IN_APPIMAGE,
